requirements.txt
boto3
python-dotenv
torch
scikit-learn
pandas
numpy
matplotlib
seaborn
flask
gunicorn
pytest
pytest-cov
locust
requests
